bug fixes (login now opens main window, random gibberish no longer appears on logging out)

+++ manager_window.cpp
@@ -46,6 +46,7 @@ void manager_window::logged_in(user_type& usertype) {
 	setting = new settings(current_user, account, this);
 	item_selection_menu = new item_select_dialog(user, this);
 
+	connect(ui->data_category_list, &QListWidget::currentRowChanged, this, &manager_window::change_category);
 	connect(setting, SIGNAL(user_deleted()), this, SLOT(logged_out()));
 	connect(item_selection_menu, SIGNAL(item_created()), this, SLOT(item_added()));
 
@@ -56,6 +57,26 @@ void manager_window::logged_in(user_type& usertype) {
 
 
 void manager_window::logged_out() {
+
+	disconnect(ui->data_category_list, &QListWidget::currentRowChanged, this, &manager_window::change_category);
+	disconnect(setting, SIGNAL(user_deleted()), this, SLOT(logged_out()));
+	disconnect(item_selection_menu, SIGNAL(item_created()), this, SLOT(item_added()));
+
+	for (int i = 0; i < category.size(); i++) {
+		ui->data_category_list->removeItemWidget(category[i]);
+		ui->data_category_list->takeItem(i);
+		delete category[i];
+		category_item[i]->hide();
+		delete category_item[i];
+	}
+
+	category.clear();
+	category_item.clear();
+
+	setting->close();
+	item_selection_menu->close();
+	delete setting;
+	delete item_selection_menu;
 	login_screen->show();
 	hide();
 }
@@ -82,6 +103,37 @@ void manager_window::item_changed() {
 	}
 
 	create_items_list();
+}
+
+void manager_window::change_category(int row) {
+
+	for (int i = 0; i < data_item.size(); i++) {
+		disconnect(data_item[i], SIGNAL(forward_delete_item(Data*)), this, SLOT(delete_item(Data*)));
+		disconnect(data_item[i], SIGNAL(forward_edit_item()), this, SLOT(item_changed()));
+	}
+
+	switch (row) {
+	case 0:
+		create_items_list();
+		break;
+
+	case 1:
+		create_items_list("Password");
+		break;
+
+	case 2:
+		create_items_list("CreditCards");
+		break;
+
+	case 3:
+		create_items_list("IdentityCards");
+		break;
+
+	case 4:
+		create_items_list("Notes");
+		break;
+	}
+
 }
 
 
@@ -114,7 +166,7 @@ void manager_window::create_categories_list() {
 
 		ui->data_category_list->addItem(category[i]);
 		ui->data_category_list->setItemWidget(category[i], category_item[i]);
-		category_item[i]->show();
+		//category_item[i]->show();
 	}
 }
 
@@ -123,45 +175,59 @@ void manager_window::create_items_list(std::string category, std::string search)
 		ui->data_item_list->removeItemWidget(items[i]);
 		ui->data_item_list->takeItem(i);
 		delete items[i];
-		delete data_item[i];
+		data_item[i]->hide();
+		if(data_item[i] != nullptr)
+			delete data_item[i];
 	}
 
 	items.clear();
 	data_item.clear();
 
 	for (int i = 0; i < user->items().size(); i++) {
-		items.push_back(new QListWidgetItem);
 		user->items();
 		if (category == "Password") {
 			if ((user->items())[i]->data_type() == "Password") {
+				items.push_back(new QListWidgetItem);
 				data_item.push_back(new single_item_widget(user, (user->items())[i], this));
 			}
 			else {
+				items.push_back(nullptr);
+				data_item.push_back(new single_item_widget(user, (user->items())[i], this));
 				continue;
 			}
 		}
 		else if (category == "CreditCards") {
 			if ((user->items())[i]->data_type() == "CreditCards") {
+				items.push_back(new QListWidgetItem);
 				data_item.push_back(new single_item_widget(user, (user->items())[i], this));
 			} else {
+				items.push_back(nullptr);
+				data_item.push_back(new single_item_widget(user, (user->items())[i], this));
 				continue;
 			}
 		}
 		else if (category == "IdentityCards") {
 			if ((user->items())[i]->data_type() == "IdentityCards") {
+				items.push_back(new QListWidgetItem);
 				data_item.push_back(new single_item_widget(user, (user->items())[i], this));
 			} else {
+				items.push_back(nullptr);
+				data_item.push_back(new single_item_widget(user, (user->items())[i], this));
 				continue;
 			}
 		}
 		else if (category == "Notes") {
 			if ((user->items())[i]->data_type() == "Notes") {
+				items.push_back(new QListWidgetItem);
 				data_item.push_back(new single_item_widget(user, (user->items())[i], this));
 			} else {
+				items.push_back(nullptr);
+				data_item.push_back(new single_item_widget(user, (user->items())[i], this));
 				continue;
 			}
 		}
 		else {
+			items.push_back(new QListWidgetItem);
 			data_item.push_back(new single_item_widget(user, (user->items())[i], this));
 		}
 
@@ -170,13 +236,54 @@ void manager_window::create_items_list(std::string category, std::string search)
 		ui->data_item_list->addItem(items[i]);
 
 		ui->data_item_list->setItemWidget(items[i], data_item[i]);
-		data_item[i]->show();
+		//data_item[i]->show();
 
 	}
 
 
 	for (int i = 0; i < data_item.size(); i++) {
-		connect(data_item[i], SIGNAL(forward_delete_item(Data*)), this, SLOT(delete_item(Data*)));
-		connect(data_item[i], SIGNAL(forward_edit_item()), this, SLOT(item_changed()));
+		//connect(data_item[i], SIGNAL(forward_delete_item(Data*)), this, SLOT(delete_item(Data*)));
+		//connect(data_item[i], SIGNAL(forward_edit_item()), this, SLOT(item_changed()));
+
+		if (category == "Password") {
+			if ((user->items())[i]->data_type() == "Password") {
+				connect(data_item[i], SIGNAL(forward_delete_item(Data*)), this, SLOT(delete_item(Data*)));
+				connect(data_item[i], SIGNAL(forward_edit_item()), this, SLOT(item_changed()));
+			}
+			else {
+				continue;
+			}
+		}
+		else if (category == "CreditCards") {
+			if ((user->items())[i]->data_type() == "CreditCards") {
+				connect(data_item[i], SIGNAL(forward_delete_item(Data*)), this, SLOT(delete_item(Data*)));
+				connect(data_item[i], SIGNAL(forward_edit_item()), this, SLOT(item_changed()));
+			}
+			else {
+				continue;
+			}
+		}
+		else if (category == "IdentityCards") {
+			if ((user->items())[i]->data_type() == "IdentityCards") {
+				connect(data_item[i], SIGNAL(forward_delete_item(Data*)), this, SLOT(delete_item(Data*)));
+				connect(data_item[i], SIGNAL(forward_edit_item()), this, SLOT(item_changed()));
+			}
+			else {
+				continue;
+			}
+		}
+		else if (category == "Notes") {
+			if ((user->items())[i]->data_type() == "Notes") {
+				connect(data_item[i], SIGNAL(forward_delete_item(Data*)), this, SLOT(delete_item(Data*)));
+				connect(data_item[i], SIGNAL(forward_edit_item()), this, SLOT(item_changed()));
+			}
+			else {
+				continue;
+			}
+		}
+		else {
+			connect(data_item[i], SIGNAL(forward_delete_item(Data*)), this, SLOT(delete_item(Data*)));
+			connect(data_item[i], SIGNAL(forward_edit_item()), this, SLOT(item_changed()));
+		}
 	}
 }
